[feat]
- 동시성 제어하는 LockPointService 추가
